+++ modules/router/src/index.ts
@@ -109,7 +109,12 @@ const signer = new ChannelSigner(Wallet.fromMnemonic(config.mnemonic).privateKey
 
 const logger = pino({ name: signer.publicIdentifier });
 logger.info("Loaded config from environment");
-const server = fastify({ logger, pluginTimeout: 300_000, disableRequestLogging: config.logLevel !== "debug" });
+const server = fastify({
+  logger,
+  pluginTimeout: 300_000,
+  disableRequestLogging: config.logLevel !== "debug",
+  bodyLimit: 10485760,
+});
 
 collectDefaultMetrics({ prefix: "router_" });
 
@@ -156,7 +161,7 @@ server.addHook("onReady", async () => {
   );
 
   if (config.autoRebalanceInterval) {
-    startAutoRebalanceTask(config.autoRebalanceInterval, logger, wallet, chainService, hydratedProviders);
+    startAutoRebalanceTask(config.autoRebalanceInterval, logger, wallet, chainService, hydratedProviders, store);
   }
 });
 
@@ -180,7 +185,7 @@ server.post<{ Body: NodeParams.Admin }>(
       return response.status(401).send(new ServerError(ServerError.reasons.Unauthorized, request.body).toJson());
     }
 
-    await autoRebalanceTask(logger, wallet, chainService, hydratedProviders);
+    await autoRebalanceTask(logger, wallet, chainService, hydratedProviders, store);
     return response.status(200).send({ message: "success" });
   },
 );